move web-bridge to domain
